add theme mode quick select in menu bar
