+++ src/client/containers/NoteMenuBar.tsx
@@ -9,24 +9,18 @@ import {
   RefreshCw,
   Loader,
   Settings,
-  Sun,
-  Moon,
   Clipboard as ClipboardCmp,
 } from 'react-feather'
 
 import { TestID } from '@resources/TestID'
 import { LastSyncedNotification } from '@/components/LastSyncedNotification'
 import { NoteItem, CategoryItem } from '@/types'
-import {
-  toggleSettingsModal,
-  togglePreviewMarkdown,
-  toggleDarkTheme,
-  updateCodeMirrorOption,
-} from '@/slices/settings'
+import { toggleSettingsModal, togglePreviewMarkdown } from '@/slices/settings'
 import { toggleFavoriteNotes, toggleTrashNotes } from '@/slices/note'
-import { getCategories, getNotes, getSync, getSettings } from '@/selectors'
+import { getCategories, getNotes, getSync } from '@/selectors'
 import { downloadNotes, isDraftNote, getShortUuid, copyToClipboard } from '@/utils/helpers'
 import { sync } from '@/slices/sync'
+import { NewThemeService } from '@/newThemeService'
 
 export const NoteMenuBar = () => {
   // ===========================================================================
@@ -36,7 +30,6 @@ export const NoteMenuBar = () => {
   const { notes, activeNoteId } = useSelector(getNotes)
   const { categories } = useSelector(getCategories)
   const { syncing, lastSynced, pendingSync } = useSelector(getSync)
-  const { darkTheme } = useSelector(getSettings)
 
   // ===========================================================================
   // Other
@@ -80,9 +73,6 @@ export const NoteMenuBar = () => {
   const _sync = (notes: NoteItem[], categories: CategoryItem[]) =>
     dispatch(sync({ notes, categories }))
   const _toggleSettingsModal = () => dispatch(toggleSettingsModal())
-  const _toggleDarkTheme = () => dispatch(toggleDarkTheme())
-  const _updateCodeMirrorOption = (key: string, value: any) =>
-    dispatch(updateCodeMirrorOption({ key, value }))
 
   // ===========================================================================
   // Handlers
@@ -93,10 +83,6 @@ export const NoteMenuBar = () => {
   const trashNoteHandler = () => _toggleTrashNotes(activeNoteId)
   const syncNotesHandler = () => _sync(notes, categories)
   const settingsHandler = () => _toggleSettingsModal()
-  const toggleDarkThemeHandler = () => {
-    _toggleDarkTheme()
-    _updateCodeMirrorOption('theme', darkTheme ? 'base16-light' : 'new-moon')
-  }
   const togglePreviewHandler = () => {
     togglePreviewIcon(!isToggled)
     _togglePreviewMarkdown()
@@ -164,10 +150,8 @@ export const NoteMenuBar = () => {
           )}
           <span className="sr-only">Sync notes</span>
         </button>
-        <button className="note-menu-bar-button" onClick={toggleDarkThemeHandler}>
-          {darkTheme ? <Sun aria-hidden="true" size={18} /> : <Moon aria-hidden="true" size={18} />}
-          <span className="sr-only">Themes</span>
-        </button>
+
+        <NewThemeService />
 
         <button className="note-menu-bar-button" onClick={settingsHandler}>
           <Settings aria-hidden="true" size={18} />
